refactor: rearrange directory structure of core
