+++ src/core/database/index.ts
@@ -1,0 +1,10 @@
+import { DatabaseManager } from './manager';
+
+export * from './manager';
+
+/**
+ * Global database manager singleton instance.
+ */
+const databaseManager = new DatabaseManager();
+
+export { databaseManager as db };
