+++ src/core/database/manager.ts
