+++ src/core/options/manager.ts
